100th commit!!!!!!!! Oh and also added gun icons
Gee, I don't know what to say... I'd like to thank all of my amazing
fanbase for their support. I'd also like to thank the rest of the team,
without them all of this would not have been possible.

An icon of the selected gun is now displayed next to ammo counter